+++ examples/tutorials/10_agentic/10_temporal/010_agent_chat/project/run_worker.py
@@ -8,6 +8,7 @@ from agentex.lib.environment_variables import EnvironmentVariables
 from workflow import At010AgentChatWorkflow
 
 
+
 environment_variables = EnvironmentVariables.refresh()
 
 logger = make_logger(__name__)