+++ src/agentex/types/shared/__init__.py
@@ -1,4 +1,3 @@
 # File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
-__title__ = "agentex"
-__version__ = "0.2.6"  # x-release-please-version
+from .delete_response import DeleteResponse as DeleteResponse